requirements.txt
langchain==0.1.11
openai==1.13.3
elevenlabs>=0.2.0
python-dotenv==1.0.1
ffmpeg-python>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# TTS dependencies
requests>=2.28.0
pydub>=0.25.1

# Image generation dependencies
openai>=1.0.0  # For DALL-E
google-cloud-aiplatform>=1.0.0  # For Vertex AI
vertexai>=0.0.1  # For Vertex AI image generation

# Video assembly dependencies
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.0.0
flake8>=6.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
oauth2client>=4.1.3
tiktok-api>=0.1.0
instagram-private-api>=1.6.0
praw==7.7.1
gspread==5.12.4
pyyaml==6.0.1 